refactor: 데이터 매핑 함수 분리
